requirements.txt
Appium-Python-Client==0.31
beautifulsoup4==4.8.2
nb-log==5.0
openpyxl==3.0.7
PyMySQL==0.10.0
PyYAML==5.3.1
requests==2.26.0
selenium==3.141.0
urllib3==1.25.11
xlrd==1.2.0
XlsxWriter==1.1.8
xlutils==2.0.0
xlwt==1.3.0
